+++ src/Interpreters/DatabaseCatalog.cpp
@@ -608,11 +608,10 @@ Dependencies DatabaseCatalog::getDependencies(const StorageID & from) const
         return {};
     return Dependencies(iter->second.begin(), iter->second.end());
 }
- 
-void DatabaseCatalog::getViewDependencies(ViewDependencies & view_dependencies_) const
+
+ViewDependencies DatabaseCatalog::getViewDependencies() const;
 {
-    std::lock_guard lock{databases_mutex};
-    view_dependencies_ = view_dependencies;
+    return view_dependencies;
 }
 
 void
