+++ src/Storages/System/StorageSystemViews.h
@@ -6,16 +6,16 @@
 namespace DB
 {
 
-class StorageSystemTableViews final : public ext::shared_ptr_helper<StorageSystemTableViews>, public IStorageSystemOneBlock<StorageSystemTableViews>
+class StorageSystemViews final : public ext::shared_ptr_helper<StorageSystemViews>, public IStorageSystemOneBlock<StorageSystemViews>
 {
-    friend struct ext::shared_ptr_helper<StorageSystemTableViews>;
+    friend struct ext::shared_ptr_helper<StorageSystemViews>;
 protected:
     using IStorageSystemOneBlock::IStorageSystemOneBlock;
 
     void fillData(MutableColumns & res_columns, const Context & context, const SelectQueryInfo & query_info) const override;
 
 public:
-    std::string getName() const override { return "TableViews"; }
+    std::string getName() const override { return "SystemViews"; }
 
     static NamesAndTypesList getNamesAndTypes();
